Move `DimensionName` to `metadata` namespace
